Added a healing ability for sorcerer
I added a healing ability for the sorcerer.  (Healing Shield - Create a forcefield of healing magic. Heals the player by 20 health per 0.5 seconds 5 times, allowing the player to heal by 100 health points in total when used.)

This ability replaces the healing potion for the sorcerer and allows them to use magic to heal instead.